src/backend/requirements.txt
redis
orjson
openai
pandas
pyarrow
fastapi
requests
python-dotenv
uvicorn[standard]
